+++ Library/include/CSP/Systems/Assets/AssetSystem.h
@@ -100,7 +100,7 @@ public:
     /// @brief Deletes a given array of asset collections.
     /// @param AssetCollections csp::common::Array<AssetCollection> : The array of asset collections to delete
     /// @param Callback NullResultCallback : callback when asynchronous task finishes
-    CSP_ASYNC_RESULT void DeleteMultipleAssetCollections(csp::common::Array<AssetCollection>& AssetCollections, NullResultCallback Callback);
+    CSP_ASYNC_RESULT void DeleteMultipleAssetCollections(const csp::common::Array<AssetCollection>& AssetCollections, NullResultCallback Callback);
 
     /// @brief Copies an array of asset collections to another space. Note that all source asset collections must belong to the same space.
     /// @param SourceAssetCollections csp::common::Array<AssetCollection> : The array of asset collections to copy. They must all belong to the same
@@ -109,7 +109,7 @@ public:
     /// @param CopyAsync const csp::common::Optional<bool> : Whether to instruct the services to perform the copy of the asset collections
     /// asynchronously.
     /// @param Callback NullResultCallback : callback when asynchronous task finishes
-    CSP_ASYNC_RESULT void CopyAssetCollectionsToSpace(csp::common::Array<AssetCollection>& SourceAssetCollections,
+    CSP_ASYNC_RESULT void CopyAssetCollectionsToSpace(const csp::common::Array<AssetCollection>& SourceAssetCollections,
         const csp::common::String& DestSpaceId, bool CopyAsync, AssetCollectionsResultCallback Callback);
 
     /// @brief Finds an asset collection by its Id.
@@ -298,7 +298,7 @@ public:
     /// @param AssetTags csp::common::Array<csp::common::String>& : Tags to be associated with the created material.
     /// @param Callback MaterialResultCallback : Callback when asynchronous task finishes.
     CSP_ASYNC_RESULT void CreateMaterial(const csp::common::String& Name, const csp::systems::EShaderType ShaderType,
-        const csp::common::String& SpaceId, csp::common::Map<csp::common::String, csp::common::String>& Metadata,
+        const csp::common::String& SpaceId, csp::common::Map<csp::common::String, csp::common::String> Metadata,
         const csp::common::Array<csp::common::String>& AssetTags, MaterialResultCallback Callback);
 
     /// @brief Updates an existing material's properties.